cmake: fix output location of import libraries and DLLs
As observed by Edward Thomson, the libgit2 DLL built by Windows will not
end up in the top-level build directory but instead inside of the 'src/'
subdirectory. While confusing at first because we are actually setting
the LIBRARY_OUTPUT_DIRECTORY to the project's binary directory, the
manual page of LIBRARY_OUTPUT_DIRECTORY clears this up:

    There are three kinds of target files that may be built: archive,
    library, and runtime. Executables are always treated as runtime
    targets. Static libraries are always treated as archive targets.
    Module libraries are always treated as library targets. For non-DLL
    platforms shared libraries are treated as library targets. For DLL
    platforms the DLL part of a shared library is treated as a runtime
    target and the corresponding import library is treated as an archive
    target. All Windows-based systems including Cygwin are DLL
    platforms.

So in fact, DLLs and import libraries are not treated as libraries at
all by CMake but instead as runtime and archive targets. To fix the
issue, we can thus simply set the variables RUNTIME_OUTPUT_DIRECTORY and
ARCHIVE_OUTPUT_DIRECTORY to the project's root binary directory.