tests: introduce ability to use config-driven ignore tests
In an ideal world, all projects aiming to implement functional aspects
of Git would share a common test suite. This test suite would be in a
generic format readable by all projects and could then be used to drive
unit tests. Like this, it can be made sure that all these different
projects would at least behave the same regarding to this common base of
tests.

One of the easiest candidates to create such a shared test suite is the
gitignore rules. All that is required is a format that specifies all the
gitignore files that should exist as well as a list of files that should
or should not be ignored, respectively. A natural choice here is to
simply use Git configuration files to convey this information, as all
Git projects should already know how to parse them. Like this, only
minimal additional code is required to have a test driver for those
specifications.

This commit implements a first proof of concept for such a test
specification. Each test specification is contained in a single file.
This file has several keys:

    - "test.name": name for the test, which can be used by the unit test
      implementation to give hints to the user which test is currently
      being executed
    - "assert.ignored": path to a file which should be ignored
    - "assert.not-ignored": path to a file which should not be ignored
    - "gitignore.path": path of a new gitignore file. All the following
      rules will be written into this path.
    - "gitignore.rule": rule that should be written to the current
      "gitignore.path" file. Can only be specified if a "gitignore.path"
      has been configured before.

Due to the dependence between "gitignore.rule" and "gitignore.path", the
configuration file needs to always be parsed from top to bottom.
Adding a second "gitignore.path" key will cause all the following
"gitignore.rule" values to instead be added to the new file path.

This is obviously only a proof of concept. First, libgit2 fails to
provide a way of iterating through all config keys from top to bottom,
but will only allow you to iterate through them in indeterminate order.
Second, we shouldn't be adding the test specifications to our own
projects but instead have them part of a submodule. Its repository
should probably live at an official location, where it is accessible to
other projects like git.git and JGit, as well.

So currently, I'm just seeking feedback in our small group and will
later on publish it for comments on the Git mailing list (obviously,
with code specific to git.git such that they can use the test specs for
their project, as well).